Update What Is Bash
My problem is that I've been trying to run what I see on the screen (which may not make sense for what I'm doing). In this case, I think the tutorial is saving their files to a location where they don't need to go searching for it (whereas I have to go to my folder). Once I go to my folder, it seems to work!